+++ source/cpp/ctlloadresources.cpp
@@ -24,7 +24,6 @@
 #include <fstream>
 #include <string>
 #include <boost/algorithm/string.hpp>
-#include "ctliface.h"
 #include "ctlloadresources.h"
 #include "ctlwindowsimpl.h"
 #include "cppfunc.h"
@@ -115,7 +114,7 @@ namespace
             lineQueue.pop();
             lineQueue.push(line);
         }
-        auto crcVal = crc32_aux(reinterpret_cast<unsigned char*>(totalBuf.data()), static_cast<unsigned int>(totalBuf.size()));
+        auto crcVal = crc32_aux(reinterpret_cast<unsigned char*>(totalBuf.data()), totalBuf.size());
         try
         {
             uint64_t crc = std::stoul(lineQueue.front());
